refactor: move KugouQrLoginActivity to auth subpackage and update references
