+++ app/src/main/java/moe/ouom/neriplayer/ui/screen/tab/SettingsScreen.kt
@@ -189,7 +189,6 @@ import moe.ouom.neriplayer.ui.viewmodel.auth.BiliAuthViewModel
 import moe.ouom.neriplayer.ui.viewmodel.auth.YouTubeAuthEvent
 import moe.ouom.neriplayer.ui.viewmodel.auth.KugouAuthViewModel
 import moe.ouom.neriplayer.ui.viewmodel.auth.YouTubeAuthViewModel
-import moe.ouom.neriplayer.activity.KugouQrLoginActivity
 import moe.ouom.neriplayer.ui.screen.tab.settings.auth.SettingsKugouAuthDialogs
 import moe.ouom.neriplayer.ui.viewmodel.debug.NeteaseAuthEvent
 import moe.ouom.neriplayer.ui.viewmodel.debug.NeteaseAuthViewModel
