+++ app/src/main/java/moe/ouom/neriplayer/ui/screen/tab/settings/auth/SettingsCookieAuthDialogs.kt
@@ -54,9 +54,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import moe.ouom.neriplayer.R
-import moe.ouom.neriplayer.activity.BiliQrLoginActivity
-import moe.ouom.neriplayer.activity.KugouQrLoginActivity
-import moe.ouom.neriplayer.activity.YouTubeWebLoginActivity
+import moe.ouom.neriplayer.activity.auth.KugouQrLoginActivity
 import moe.ouom.neriplayer.core.di.AppContainer
 import moe.ouom.neriplayer.ui.component.sheet.bottomSheetDragBlocker
 import moe.ouom.neriplayer.ui.screen.tab.settings.component.InlineMessage
